# docs/03-openapi.yaml
openapi: 3.1.0
info:
  title: API Signup/Login
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: |
    API REST de ejemplo para registro y autenticación: registro y login con email/contraseña,
    login con Google (OIDC — GIS token flow), refresh con rotación, logout, acceso a un recurso
    protegido, acceso por magic link sin contraseña, cambio de contraseña y recuperación de
    contraseña por email.

    **Contrato de la Fase 2 de Spec Driven Development** — derivado de las historias de usuario
    (`01-historias-de-usuario.md`, US-01 a US-12) y de las consideraciones técnicas
    (`00-consideraciones-tecnicas.md`). La representación gráfica de los flujos está en
    `02-flujos-registro-autenticacion.md`.

    Reglas transversales del contrato:

    - Envelope único: éxito `{ "data": ... }`; error `{ "error": { code, message, details?, requestId } }`.
    - `400` = JSON/header malformado · `422` = validación de negocio (zod, con `details` por campo) ·
      `409` = conflicto de identidad (nunca auto-linking) · `429` = rate limit (con `Retry-After`) ·
      `404`/`405` = ruta/verbo inexistentes · `500` = error interno genérico con `requestId`.
    - `401` de autenticación: mensajes genéricos e idénticos (anti-enumeración); jamás se revela si un
      email existe.
    - Access token: JWT HS256 con vigencia de 5-15 min; el algoritmo se fija en el servidor (nunca se
      confía en el header `alg`).
...
servers:
  - url: http://localhost:3000/api/v1
...
security: []
paths:
...
  /auth/magic-link/request:
    post:
      tags: [Auth]
      operationId: requestMagicLink
      summary: Solicitar magic link (sign in sin contraseña)
      description: |
        Genera y envía un enlace de acceso por email (US-09). La respuesta `200 { data: { ok: true } }`
        es **idéntica** exista o no el email, y se realiza la misma cantidad de trabajo (generar token,
        persistir su hash, enviar email) en ambos casos — anti-enumeración estricta: ni la respuesta ni
        un side-channel temporal revelan si la cuenta está registrada. El servidor persiste solo el
        **hash SHA-256** del token opaco (≥ 32 bytes aleatorios), nunca el token en claro. El enlace
        expira en un TTL corto configurable (`MAGIC_LINK_TTL_MINUTES`, default 15, máx 60). La auto-cuenta
        se resuelve en el consume (US-10 AC-02): un link enviado a un email no registrado crea la cuenta.
        El campo `intent` selecciona el canal (US-11/12): `login` (default) envía la URL de consumo de
        sesión; `password_reset` envía la URL de recuperación de contraseña con la MISMA forma de respuesta
        y la MISMA cantidad de trabajo (anti-enumeración también en el canal de recuperación).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MagicLinkRequest'
            examples:
              valido:
                summary: Payload válido (intent omiso = login, US-09)
                value:
                  email: ana@example.com
              reseteo:
                summary: Solicitud de recuperación de contraseña (US-12)
                value:
                  email: ana@example.com
                  intent: password_reset
      responses:
        '200':
          description: Enlace enviado (o no) — respuesta idéntica siempre (anti-enumeración).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MagicLinkRequestResponse'
              examples:
                ok:
                  summary: Solicitud aceptada (US-09 AC-01/AC-02)
                  value:
                    data:
                      ok: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/magic-link/consume:
    post:
      tags: [Auth]
      operationId: consumeMagicLink
      summary: Consumir magic link (login por enlace)
      description: |
        Intercambia un token de magic link válido y vigente por el par de tokens (US-10). La respuesta
        tiene la misma forma que `/login` (el cliente no distingue el proveedor). **Auto-cuenta**: si el
        email del link no está registrado, se crea el usuario (`provider = 'magic'`, `email_verified = 1`)
        y se responde 200. Si ya está registrado (local/Google), se marca `email_verified = 1` y se emite
        sesión sobre la cuenta existente. **Un solo uso**: el token se marca `used` al consumirse. Token
        inexistente, revocado o vencido → 401 `MAGIC_LINK_INVALID`, idéntico en forma (anti-enumeración).
        El consumo de sesión solo acepta enlaces `purpose = 'login'`: un enlace de recuperación
        (`password_reset`) presentado aquí responde el mismo 401 idéntico (US-12 F3).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MagicLinkConsumeRequest'
            examples:
              valido:
                summary: Payload válido
                value:
                  token: ejOpc0aQnFgW4s9XyZ2bCdEfGhIjKlMnOpQrStUvWx
      responses:
        '200':
          description: Sesión iniciada; mismo contrato que /login / /auth/google.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                ok:
                  summary: Acceso por enlace (US-10 AC-01/AC-02/AC-03)
                  value:
                    data:
                      accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI5ZjBjN2EyZS00YjZkLTRjOGEtOWYyZS0zYTFiMmMzZDRlNWYiLCJpc3MiOiJhcGktc2lnbnVwLWxvZ2luIiwiYXVkIjoiYXBpLWNsaWVudCIsImlhdCI6MTc4Nzg2OTA3MCwiZXhwIjoxNzg3ODcwNDcwLCJuYmYiOjE3ODc4NjkwNzAsImp0aSI6ImEwMWI1Y2NjIn0.ejemplo_firma
                      refreshToken: v4.local.ejemplo-opaco-solo-server
                      user:
                        id: 9f0c7a2e-4b6d-4c8a-9f2e-3a1b2c3d4e5f
                        email: ana@example.com
                        createdAt: '2026-08-27T10:15:30.000Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Token inexistente, revocado o vencido — forma idéntica (anti-enumeración).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /auth/password/reset:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: Recuperar contraseña (consumo del magic link de reset)
      description: |
        Consume un magic link emitido con `intent: password_reset` y reemplaza (o asigna) la
        contraseña del email del enlace (US-12). El token ES la credencial: la ruta solo lleva
        `authLimiter` (sin requireAuth). **NO emite sesión** — respuesta `204`; el cliente redirige
        al login con el secreto nuevo. **F2**: email aún no registrado → auto-cuenta local (el enlace
        prueba la posesión del email, coherente con US-10 AC-02). **F1**: el reset revoca TODAS las
        sesiones del usuario. Token inexistente, vencido, ya usado o de otro propósito (`login`) →
        401 `MAGIC_LINK_INVALID`, idéntico en forma (anti-enumeración).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
            examples:
              valido:
                summary: Reset con token del email (US-12 AC-01)
                value:
                  token: ejOpc0aQnFgW4s9XyZ2bCdEfGhIjKlMnOpQrStUvWx
                  password: nuevaContraseñaSegura789
      responses:
        '204':
          description: Contraseña reemplazada/asignada y sesiones revocadas; sin sesión emitida (US-12 AC-01/AC-02).
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Token inexistente, vencido, ya usado o de otro propósito — forma idéntica (anti-enumeración).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token JWT HS256 (5-15 min). El algoritmo de verificación se fija en el servidor.
  schemas:
    Email:
      type: string
      format: email
      maxLength: 254
      description: Email normalizado a minúsculas (Value Object del dominio, sección 6.1).
      example: ana@example.com
    Password:
      type: string
      minLength: 8
      maxLength: 64
      description: NIST 800-63B — por longitud, sin reglas de complejidad. Solo hash argon2id en reposo.
      example: contraseñaSegura123
...
    PasswordResetRequest:
      type: object
      required: [token, password]
      additionalProperties: false
      properties:
        token:
          $ref: '#/components/schemas/MagicLinkToken'
        password:
          $ref: '#/components/schemas/Password'
...
    MagicLinkIntent:
      type: string
      enum: [login, password_reset]
      description: |
        Canal del magic link (US-09/10 vs US-12). `login` (default) → enlace de acceso con sesión;
        `password_reset` → enlace de recuperación de contraseña (misma forma de respuesta y mismo
        trabajo realizado — anti-enumeración también en el reset). Se persiste en `magic_links.purpose`.
      example: login
    MagicLinkRequest:
      type: object
      required: [email]
      additionalProperties: false
      properties:
        email:
          $ref: '#/components/schemas/Email'
        intent:
          $ref: '#/components/schemas/MagicLinkIntent'
    MagicLinkRequestResponse:
      type: object
      required: [data]
      additionalProperties: false
      properties:
        data:
          type: object
          required: [ok]
          additionalProperties: false
          properties:
            ok:
              type: boolean
              description: Siempre true — respuesta idéntica exista o no el email (anti-enumeración, US-09 AC-02).
              example: true
    MagicLinkToken:
      type: string
      description: |
        Token opaco aleatorio (≥ 32 bytes, base64url) que viaja en la URL de consumo `?token=<opaco>`.
        El servidor persiste solo su hash SHA-256. Un solo uso; TTL corto configurable (US-09 AC-03/AC-04).
      example: ejOpc0aQnFgW4s9XyZ2bCdEfGhIjKlMnOpQrStUvWx
    MagicLinkConsumeRequest:
      type: object
      required: [token]
      additionalProperties: false
      properties:
        token:
          $ref: '#/components/schemas/MagicLinkToken'
    AccessToken:
      type: string
      description: |
        JWT HS256 del proyecto. Claims: sub (users.id), iss, aud, exp, iat, nbf y jti.
        Vigencia 5-15 min. Algoritmo fijado en el servidor (jamás del header alg).
      example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.ejemplo.ejemplo_firma
    RefreshToken:
      type: string
      description: |
        Token opaco (no-JWT) con vigencia de 7-30 días. El servidor guarda solo su hash SHA-256 + jti;
        se rota en cada refresh y el reuso revoca toda la familia.
      example: v4.local.ejemplo-opaco-solo-server
    UserPublic:
      type: object
      required: [id, email, createdAt]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
          description: UUID v4 de users.id (nunca una clave autoincremental).
          example: 9f0c7a2e-4b6d-4c8a-9f2e-3a1b2c3d4e5f
        email:
          $ref: '#/components/schemas/Email'
        createdAt:
          type: string
          format: date-time
          description: Fecha de alta del usuario.
          example: '2026-08-27T10:15:30.000Z'
    AuthResponse:
      type: object
      required: [data]
      additionalProperties: false
      properties:
        data:
          type: object
          required: [accessToken, refreshToken, user]
          additionalProperties: false
          properties:
            accessToken:
              $ref: '#/components/schemas/AccessToken'
            refreshToken:
              $ref: '#/components/schemas/RefreshToken'
            user:
              $ref: '#/components/schemas/UserPublic'
...
    ValidationIssue:
      type: object
      required: [field, issue]
      additionalProperties: false
      properties:
        field:
          type: string
          description: Campo del payload que falló la validación.
          example: email
        issue:
          type: string
          description: Identificador estable del problema (zod).
          example: invalid_email
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message, requestId]
          additionalProperties: false
          properties:
            code:
              type: string
              description: |
                Identificador estable y programable. Complemento del catálogo de US-06: los códigos
                documentados son VALIDATION_ERROR, INVALID_CREDENTIALS, EMAIL_ALREADY_EXISTS,
                ACCOUNT_EXISTS_WITH_GOOGLE, EMAIL_NOT_VERIFIED, MAGIC_LINK_INVALID,
                ACCOUNT_HAS_NO_PASSWORD y RATE_LIMITED; UNAUTHORIZED y MALFORMED_REQUEST completan la
                matriz 401/400; NOT_FOUND y METHOD_NOT_ALLOWED cubren el 404/405 centralizado
                (doc 00 → ítem 24); INTERNAL_ERROR para 500.
              enum:
                - VALIDATION_ERROR
                - INVALID_CREDENTIALS
                - EMAIL_ALREADY_EXISTS
                - ACCOUNT_EXISTS_WITH_GOOGLE
                - EMAIL_NOT_VERIFIED
                - MAGIC_LINK_INVALID
                - ACCOUNT_HAS_NO_PASSWORD
                - RATE_LIMITED
                - UNAUTHORIZED
                - MALFORMED_REQUEST
                - NOT_FOUND
                - METHOD_NOT_ALLOWED
                - INTERNAL_ERROR
            message:
              type: string
              description: Mensaje legible por humanos; genérico e idéntico en 401 (anti-enumeración).
            details:
              type: array
              description: Opcional; estructura los errores de validación por campo y el proveedor sugerido en 409.
              items:
                $ref: '#/components/schemas/ValidationIssue'
            requestId:
              type: string
              description: Correlación con los logs estructurados (pino-http, doc 00 → ítem 11).
              example: req_01J4X8K2YQ
  responses:
    BadRequest:
      description: JSON/body malformado o header inválido — envelope sin detalles internos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            malformado:
              value:
                error:
                  code: MALFORMED_REQUEST
                  message: El cuerpo de la petición no es JSON válido.
                  requestId: req_01J4X8K2YW
...
    ValidationError:
      description: Falla de validación de negocio (zod) — details por campo, preceden a cualquier verificación.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            campos-invalidos:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: La petición no cumple las reglas de validación.
                  details:
                    - field: email
                      issue: invalid_email
                    - field: password
                      issue: too_short
                  requestId: req_01J4X8K2YZ
    RateLimited:
      description: Rate limit excedido (registro/login/refresh/google; por IP, y también por email en login).
      headers:
        Retry-After:
          description: Segundos antes de poder reintentar.
          schema:
            type: integer
            format: int32
            minimum: 0
          example: 60
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            excedido:
              value:
                error:
                  code: RATE_LIMITED
                  message: Demasiados intentos. Reintentá más tarde.
                  requestId: req_01J4X8K2Z0
    InternalError:
      description: Error inesperado o de infraestructura — jamás stack traces, rutas internas ni valores sensibles (doc 00 → ítem 20).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            interno:
              value:
                error:
                  code: INTERNAL_ERROR
                  message: Error interno del servidor.
                  requestId: req_01J4X8K2Z1